Moves watchImmutableRdxState function from utils.js directly into messaging-agent.js
