+++ webofneeds/won-owner-webapp/src/main/webapp/app/utils.js
@@ -60,46 +60,6 @@ export function delay(milliseconds) {
 }
 
 /**
- * `subscribe`s and watches the output of `select` for changes,
- * calling `callback` if those happen.
- * @param subscribe {function} used to subscribe
- * @param select {function} a clojure that's called to get the
- *                          value to be watched
- * @param callback {function}
- * @return {function} the unsubscribe function generated by `subscribe`
- */
-function watch(subscribe, select, callback) {
-  let unsubscribe = null;
-
-  /*
-     * creating this function (and instantly executing it)
-     * allows attaching individual previousValue to it
-     */
-  (function() {
-    let previousValue = select();
-    unsubscribe = subscribe(() => {
-      const currentValue = select();
-      if (currentValue !== previousValue) callback(currentValue, previousValue);
-      previousValue = currentValue;
-    });
-  })();
-
-  return unsubscribe;
-}
-
-/**
- * An oppinioned constiant of the generic watch that
- * for usage with redux-stores containing immutablejs-objects
- * @param redux {object} should provide `.subscribe` and `.getState`
- *                       (with the latter yielding an immutablejs-object)
- * @param path {array} an array of strings for usage with store.getIn
- * @param callback
- */
-export function watchImmutableRdxState(redux, path, callback) {
-  return watch(redux.subscribe, () => getIn(redux.getState(), path), callback);
-}
-
-/**
  * Generate string of [a-z0-9] with specified length
  * @param length
  * @returns {*}
